requirements.txt
numpy
matplotlib
opencv-python
Pillow
pyrealsense2
torch
transformers
gradio
requests
